pyproject.toml
[project]
name = "maatest"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "json-with-comments>=1.2.10",
    "loguru>=0.7.3",
    "maafw>=5.2.6",
]
